rename license config files
