+++ config/licenses/licenses.js
