+++ services/licenseStore.js
